Changed data type.

+++ Section7/BankAccount.java
@@ -6,7 +6,7 @@
 //Creating a public class.
 public class BankAccount {
     //Creating private variables.
-    private static float balance;
+    private static double balance;
     private static float withdrawal;
     private static float deposit;
     //Creating a public method.
